backend/requirements.txt
flask
flask_cors
qrcode[pil]
pillow
gunicorn